Add files via upload
The program is straightforward to use
Just open the .exe file No installation needed.
using python2exe on the code will provide with the same result as the .exe (So you can be sure there is no virus in it.

-About-
The program download and install the WGET for downloading the files.
Creates a folder in C named IT (You can find it at C:/IT)
Choose from existing categories what you want to do. 
Colors explanation
Gray = Button
Red = Can't interact with it
Green = Can interact with it (Inputs,Checkboxes,RadioButtons)
PC Specifications: Check the specifications on the machine Assistant run or do a speedtest. 
Install Software: Install software on the current machine you may need
Windows Utilities: Choose from advanced features

(P.S. That's not the finished product)